+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version [0.3.0](https://github.com/cedx/dapper.net/compare/v0.2.0...v0.3.0)
-- TODO !
+- Added the `SqlMapperExtensions` class.
 
 ## Version [0.2.0](https://github.com/cedx/dapper.net/compare/v0.1.2...v0.2.0)
 - Added the `UriTypeHandler` type handler.