+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version [0.5.0](https://github.com/cedx/dapper.net/compare/v0.4.0...v0.5.0)
+- Added the `StringEnumTypeHandler` type handler.
+- Added the `mapToIPv6` parameter to the `IPAddressTypeHandler` constructor.
+
 ## Version [0.4.0](https://github.com/cedx/dapper.net/compare/v0.3.0...v0.4.0)
 - Added the `SqlMapperExtensions.Delete(T entity)` overload.
 - Added the `SqlMapperExtensions.Insert()` method.